WIRE2JSON support
I've written the write methods for the WIRE JSON writing, with partial exception to the Transaction Detail. Currently there's support for Funds Type 0, 1, 2, S, V and Z. I am currently working on D as it is more variable than the others. I've added better catches to the variable fields in GroupHeader and AcctID. I've added checks to the JSON writer to check for null values so the additional fields added to the V and S FundsType don't get added to something like 0 or 1. There's a debug mode option if you'd like to try the horrible XML writer.